reworking structure of streams depending on functionality
