+++ airbyte-integrations/connectors/source-snowflake/source_snowflake/streams/check_connection.py
@@ -1,0 +1,72 @@
+from typing import Any, Iterable, List, Mapping, MutableMapping, Optional, Union
+
+import requests
+
+from source_snowflake.streams.snowflake_parent_stream import SnowflakeStream
+
+
+class CheckConnectionStream(SnowflakeStream):
+
+    def __init__(self, url_base, config, **kwargs):
+        super().__init__(**kwargs)
+        self._url_base = url_base
+        self.config = config
+
+    def path(
+            self, stream_state: Mapping[str, Any] = None, stream_slice: Mapping[str, Any] = None, next_page_token: Mapping[str, Any] = None
+    ) -> str:
+        """
+            path of request
+        """
+
+        return f"{self.url_base}/{self.url_suffix}"
+
+    @property
+    def url_base(self):
+        return self._url_base
+
+    @property
+    def statement(self):
+
+        """
+        Assumptions:
+            if we can see the table when showing schema, we assume we have access to the table (stream)
+            We don't need to request the table in order to make sure it is working properly
+            SHOW TABLES IN DATABASE statement does not include "system tables" in the dataset0
+            if schema is provided by the use we replace the search of tables (streams) in database by search in shema
+
+        TODO: Validate that the streams in the pushdown filter configuration are available
+        """
+        database = self.config["database"]
+        schema = self.config.get('schema', "")
+        if not schema:
+            return f"SHOW TABLES IN DATABASE {database}"
+
+        return f"SHOW TABLES IN SCHEMA {database}.{schema}"
+
+    def request_body_json(
+            self,
+            stream_state: Optional[Mapping[str, Any]],
+            stream_slice: Optional[Mapping[str, Any]] = None,
+            next_page_token: Optional[Mapping[str, Any]] = None,
+    ) -> Optional[Mapping[str, Any]]:
+        json_payload = {
+            "statement": self.statement,
+            "role": self.config['role'],
+            "warehouse": self.config['warehouse'],
+            "database": self.config['database'],
+            "timeout": "1000",
+        }
+        schema = self.config.get('schema', '')
+        if schema:
+            json_payload['schema'] = schema
+        return json_payload
+
+    def parse_response(self, response: requests.Response, **kwargs) -> Iterable[Mapping]:
+        """
+        :return an iterable containing each record in the response
+        """
+        response_json = response.json()
+        records = response_json.get("data", [])
+        yield from records
+
