+++ airbyte-integrations/connectors/source-snowflake/source_snowflake/streams/snowflake_parent_stream.py
@@ -1,0 +1,89 @@
+import logging
+import uuid
+from abc import ABC
+from typing import Any, Iterable, List, Mapping, MutableMapping, Optional, Union
+
+import requests
+from airbyte_cdk.sources.streams.http import HttpStream
+
+
+class SnowflakeStream(HttpStream, ABC):
+    """
+    TODO remove this comment
+
+    This class represents a stream output by the connector.
+    This is an abstract base class meant to contain all the common functionality at the API level e.g: the API base URL, pagination strategy,
+    parsing responses etc..
+
+    Each stream should extend this class (or another abstract subclass of it) to specify behavior unique to that stream.
+
+    Typically for REST APIs each stream corresponds to a resource in the API. For example if the API
+    contains the endpoints
+        - GET v1/customers
+        - GET v1/employees
+
+    then you should have three classes:
+    `class SnowflakeStream(HttpStream, ABC)` which is the current class
+
+    If some streams implement incremental sync, it is typical to create another class
+    `class IncrementalSnowflakeStream((SnowflakeStream), ABC)`
+
+    See the reference docs for the full list of configurable options.
+    """
+    url_suffix = "api/v2/statements"
+    url_base = ""
+
+    @property
+    def statement(self):
+        return None
+
+    def next_page_token(self, response: requests.Response) -> Optional[Mapping[str, Any]]:
+        return None
+
+    def request_params(
+            self, stream_state: Mapping[str, Any], stream_slice: Mapping[str, any] = None, next_page_token: Mapping[str, Any] = None
+    ) -> MutableMapping[str, Any]:
+        """
+        TODO: Override this method to define any query parameters to be set. Remove this method if you don't need to define request params.
+        Usually contains common params e.g. pagination size etc.
+        """
+        params = {
+            "requestId": str(uuid.uuid4()),
+            "async": "false"
+        }
+        return params
+
+    @property
+    def http_method(self) -> str:
+        return "POST"
+
+    def request_headers(
+            self,
+            stream_state: Optional[Mapping[str, Any]],
+            stream_slice: Optional[Mapping[str, Any]] = None,
+            next_page_token: Optional[Mapping[str, Any]] = None,
+    ) -> Mapping[str, Any]:
+        headers = {
+            'User-Agent': 'Airbyte',
+            'X-Snowflake-Authorization-Token-Type': 'KEYPAIR_JWT',  # to be changed when authentication method is set
+            'Content-Type': 'application/json',
+            'Accept': 'application/json',
+        }
+
+        return headers
+
+    def parse_response(self, response: requests.Response, **kwargs) -> Iterable[Mapping]:
+        """
+        :return an iterable containing each record in the response
+        """
+        response_json = response.json()
+        yield from response_json
+
+    @property
+    def primary_key(self) -> Optional[Union[str, List[str], List[List[str]]]]:
+        """
+        :return: string if single primary key, list of strings if composite primary key, list of list of strings if composite primary key consisting of nested fields.
+          If the stream has no primary keys, return None.
+        """
+        return None
+
