+++ airbyte-integrations/connectors/source-snowflake/source_snowflake/streams/util_streams.py
@@ -1,0 +1,155 @@
+from typing import Any, Iterable, List, Mapping, MutableMapping, Optional, Union
+
+import requests
+
+from source_snowflake.streams.snowflake_parent_stream import SnowflakeStream
+
+
+class TableCatalogStream(SnowflakeStream):
+    DATABASE_NAME_COLUMN = "name"
+    SCHEMA_NAME_COLUMN = "schema_name"
+
+    def __init__(self, url_base, config, **kwargs):
+        super().__init__(**kwargs)
+        self._url_base = url_base
+        self.config = config
+
+    def path(
+            self, stream_state: Mapping[str, Any] = None, stream_slice: Mapping[str, Any] = None, next_page_token: Mapping[str, Any] = None
+    ) -> str:
+        """
+            path of request
+        """
+
+        return f"{self.url_base}/{self.url_suffix}"
+
+    @property
+    def url_base(self):
+        return self._url_base
+
+    @property
+    def statement(self):
+        database = self.config["database"]
+        schema = self.config.get('schema', '')
+        if not schema:
+            return f"SHOW TABLES IN DATABASE {database}"
+
+        return f"SHOW TABLES IN SCHEMA {database}.{schema}"
+
+    def request_body_json(
+            self,
+            stream_state: Optional[Mapping[str, Any]],
+            stream_slice: Optional[Mapping[str, Any]] = None,
+            next_page_token: Optional[Mapping[str, Any]] = None,
+    ) -> Optional[Mapping[str, Any]]:
+        json_payload = {
+            "statement": self.statement,
+            "role": self.config['role'],
+            "warehouse": self.config['warehouse'],
+            "database": self.config['database'],
+            "timeout": "1000",
+        }
+        schema = self.config.get('schema', '')
+        if schema:
+            json_payload['schema'] = schema
+
+        return json_payload
+
+    @classmethod
+    def get_index_of_columns_from_names(cls, metadata_object: Mapping[Any, any], column_names: Iterable[str]) -> Mapping[str, Any]:
+        mapping_column_name_to_index = {column_name: -1 for column_name in column_names}
+        for current_index, column_object in enumerate(metadata_object["resultSetMetaData"]["rowType"]):
+            for column_name in mapping_column_name_to_index:
+                if column_object['name'] == column_name:
+                    mapping_column_name_to_index[column_name] = current_index
+
+        column_name_index_updated_filter = [0 if key_word_index == -1 else 1 for key_word_index in mapping_column_name_to_index.values()]
+
+        if not all(column_name_index_updated_filter):
+            raise ValueError('At least one index of column names is not updated. The error might be a wrong key word '
+                             'or a change in the naming of keys in resultSetMetaData of Snowflake API.\n'
+                             'To resolve this issue, compare the column name provided with keys of resultSetMetaData of Snowflake API '
+                             'and update your column names.\n'
+                             'For example, for class TableCatalogStream, compare TableCatalogStream.DATABASE_NAME_COLUMN '
+                             'and TableCatalogStream.SCHEMA_NAME_COLUMN with the keys representing this variables in resultSetMetaData '
+                             'present in the Snowflake API response')
+
+        return mapping_column_name_to_index
+
+    def parse_response(self, response: requests.Response, **kwargs) -> Iterable[Mapping]:
+        """
+        :return an iterable containing each record in the response
+        """
+        response_json = response.json()
+        column_names_to_be_extracted_from_records = [self.DATABASE_NAME_COLUMN, self.SCHEMA_NAME_COLUMN]
+        index_of_columns_from_names = self.get_index_of_columns_from_names(response_json, column_names_to_be_extracted_from_records)
+
+        database_name_index = index_of_columns_from_names[self.DATABASE_NAME_COLUMN]
+        schema_name_index = index_of_columns_from_names[self.SCHEMA_NAME_COLUMN]
+
+        for record in response_json.get("data", []):
+            yield {'schema': record[schema_name_index],
+                   'table': record[database_name_index]}
+
+
+class TableSchemaStream(SnowflakeStream):
+    def __init__(self, url_base, config, table_object, **kwargs):
+        super().__init__(**kwargs)
+        self._url_base = url_base
+        self.config = config
+        self.table_object = table_object
+
+    def path(
+            self, stream_state: Mapping[str, Any] = None, stream_slice: Mapping[str, Any] = None, next_page_token: Mapping[str, Any] = None
+    ) -> str:
+        """
+            path of request
+        """
+
+        return f"{self.url_base}/{self.url_suffix}"
+
+    @property
+    def url_base(self):
+        return self._url_base
+
+    @property
+    def statement(self):
+        database = self.config["database"]
+        schema = self.table_object["schema"]
+        table = self.table_object["table"]
+
+        return f'SELECT TOP 1 * FROM "{database}"."{schema}"."{table}"'
+
+    def request_body_json(
+            self,
+            stream_state: Optional[Mapping[str, Any]],
+            stream_slice: Optional[Mapping[str, Any]] = None,
+            next_page_token: Optional[Mapping[str, Any]] = None,
+    ) -> Optional[Mapping[str, Any]]:
+        json_payload = {
+            "statement": self.statement,
+            "role": self.config['role'],
+            "warehouse": self.config['warehouse'],
+            "database": self.config['database'],
+            "timeout": "1000",
+        }
+
+        schema = self.table_object.get('schema', '')
+        if schema:
+            json_payload['schema'] = schema
+        return json_payload
+
+    def parse_response(self, response: requests.Response, **kwargs) -> Iterable[Mapping]:
+        """
+        :return an iterable containing each record in the response
+        """
+        response_json = response.json()
+        # checks in the response nested fields response -> resultSetMetaData -> rowType
+        for row_type in response_json.get('resultSetMetaData', {'rowType': []}).get('rowType', []):
+            yield {'column_name': row_type['name'],
+                   'type': row_type['type'],
+                   }
+
+    def __str__(self):
+        return f"Current stream has this table object as constructor {self.table_object}"
+
